+++ packages/gamut/src/Modal/styles.module.scss
