+++ packages/gamut/src/index.tsx
@@ -32,8 +32,8 @@ export * from './Layout';
 export * from './Markdown';
 export * from './Menus/MenuItem';
 export * from './Menus/SideMenu';
+export * from './Dialog';
 export * from './Modal';
-export * from './ModalDeprecated';
 export * from './NotificationList';
 export * from './NotificationList/NotificationIcon';
 export * from './NotificationList/NotificationItem';
