+++ packages/styleguide/.storybook/getStories.ts
@@ -2,13 +2,7 @@ const glob = require('glob');
 const { intersection } = require('lodash');
 
 const DEFAULT_STORIES_GLOB = '../stories/**/*.stories.@(mdx|tsx)';
-const EXCLUDED_STORIES = [
-  'Truncate',
-  'Popover',
-  'Coachmark',
-  'Modal',
-  'ModalDeprecated',
-];
+const EXCLUDED_STORIES = ['Truncate', 'Popover', 'Coachmark', 'Modal', 'Modal'];
 
 const splitToChunks = (array: string[], parts: number) => {
   const copiedArray = array.concat([]);
