+++ packages/styleguide/stories/Molecules/ModalDeprecated.stories.mdx
@@ -1,18 +1,14 @@
-import {
-  ButtonDeprecated,
-  ModalDeprecated,
-  Overlay,
-} from '@codecademy/gamut/src';
+import { ButtonDeprecated, Modal, Overlay } from '@codecademy/gamut/src';
 import title from '@codecademy/macros/lib/title.macro';
 import { Canvas, Meta, Props, Story } from '@storybook/addon-docs/dist/blocks';
 import { useState } from 'react';
 
 <Meta
   title={title}
-  component={ModalDeprecated}
+  component={Modal}
   subcomponents={{ Overlay }}
   parameters={{
-    subtitle: `Simple ModalDeprecated portal to escape current DOM Context.`,
+    subtitle: `Simple Modal portal to escape current DOM Context.`,
     status: 'volatile',
     source: 'gamut',
     figmaId: 'yt7NdZTETDRPUn0CtRu3v0',
@@ -20,7 +16,7 @@ import { useState } from 'react';
 />
 
 <Canvas>
-  <Story name="ModalDeprecated">
+  <Story name="Modal">
     {(args) => {
       const [isOpen, setIsOpen] = useState(false);
       return (
@@ -28,19 +24,16 @@ import { useState } from 'react';
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <ModalDeprecated
-            isOpen={isOpen}
-            onRequestClose={() => setIsOpen(false)}
-          >
-            Close the ModalDeprecated!
-          </ModalDeprecated>
+          <Modal isOpen={isOpen} onRequestClose={() => setIsOpen(false)}>
+            Close the Modal!
+          </Modal>
         </>
       );
     }}
   </Story>
 </Canvas>
 
-<Props components={{ ModalDeprecated, Overlay }} />
+<Props components={{ Modal, Overlay }} />
 
 ### Creating a custom close button
 
@@ -48,7 +41,7 @@ By default, the Modal will render a close button.
 
 You can disable the default close button with the `hideDefaultCloseButton` prop.
 
-You can then use the `isOpen` prop to create other controls that toggle the ModalDeprecated on and off.
+You can then use the `isOpen` prop to create other controls that toggle the Modal on and off.
 
 <Canvas>
   <Story name="Without Close Default">
@@ -59,16 +52,16 @@ You can then use the `isOpen` prop to create other controls that toggle the Moda
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <ModalDeprecated
+          <Modal
             isOpen={isOpen}
             hideDefaultCloseButton
             onRequestClose={() => setIsOpen(false)}
           >
-            <p>Close the ModalDeprecated...</p>
+            <p>Close the Modal...</p>
             <ButtonDeprecated block onClick={() => setIsOpen(false)}>
               with this button instead
             </ButtonDeprecated>
-          </ModalDeprecated>
+          </Modal>
         </>
       );
     }}
@@ -79,7 +72,7 @@ You can then use the `isOpen` prop to create other controls that toggle the Moda
 
 Other than creating buttons that can close the Modal, there are other ways a Mdal can be closed by the user.
 
-1. Clicking the content outside of the ModalDeprecated
+1. Clicking the content outside of the Modal
 2. Clicking the escape key on their keyboard
 
 All of these methods default to true for accessibility reasons, and rely on setting the required `onRequestClose` prop and making sure it toggles the `isOpen` prop on the Modal.
@@ -93,7 +86,7 @@ All of these methods default to true for accessibility reasons, and rely on sett
           <ButtonDeprecated onClick={() => setIsOpen(true)}>
             Open
           </ButtonDeprecated>
-          <ModalDeprecated
+          <Modal
             isOpen={isOpen}
             hideDefaultCloseButton
             onRequestClose={() => setIsOpen(false)}
@@ -102,7 +95,7 @@ All of these methods default to true for accessibility reasons, and rely on sett
               Close this Modal by clicking the page outside of the Modal
               content, or by clicking the escape key.
             </p>
-          </ModalDeprecated>
+          </Modal>
         </>
       );
     }}
